Buildings updates
Buildings now use ingredients:
A BuildingRecipes node contains a dictionary of recipe-dictionaries and functions to return requirements and consume materials via interaction with the PlayerResources singleton.
The InputController will use BuildingRecipes to check if the requirements are satisfied. If the requirements are satisfied, the building is instanced and the materials are removed from the PlayerResources. Otherwise, the building mode is cancelled.